requirements.txt
google-genai>=0.3.0
python-dotenv>=1.0.0
gradio>=4.0.0
